Using request queues
This should improove the reliability of the extension. However I am not
yet happy, how "important" requests are handled. They will have a falg
priorize. However in order to it to be helpful, they also need to be
sent enought. Since the extension currently offers an option the set the
interval for "not priorized" requests, I have to think of a way to deal
with both of them in a single queue.